Moving ProjectDisplay component
