+++ src/components/ProjectDisplay/ProjectDisplay.svelte
